+++ devops_data_generator/adapters/gitlab/__init__.py
@@ -1,3 +1,4 @@
 from .adapter import GitLabAdapter
+from .deploy_adapter import GitLabDeployAdapter
 
-__all__ = ["GitLabAdapter"]
+__all__ = ["GitLabAdapter", "GitLabDeployAdapter"]